services/cost-scanner/requirements.txt
boto3==1.28.25
psycopg2-binary==2.9.7
python-dotenv==1.0.0
schedule==1.2.0
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.22
